requirements.txt
pandas 
langchain-groq 
langchain-community
langchain
python-dotenv
huggingface-hub
faiss-cpu
fastapi
python-multipart
pydantic
